Rename to snapshot
