--- roktux/src/test/java/com/rokt/roktux/screenshot/ScreenshotTest.kt
@@ -1,3 +1,0 @@
-package com.rokt.roktux.screenshot
-
-interface ScreenshotTest
