+++ roktux/src/test/java/com/rokt/roktux/snapshot/BasicSnapshotTest.kt
@@ -1,4 +1,4 @@
-package com.rokt.roktux.screenshot
+package com.rokt.roktux.snapshot
 
 import android.graphics.drawable.ColorDrawable
 import android.os.Looper
@@ -25,13 +25,13 @@ import org.robolectric.annotation.Config
 import org.robolectric.annotation.GraphicsMode
 
 @RunWith(AndroidJUnit4::class)
-@Category(ScreenshotTest::class)
+@Category(SnapshotTest::class)
 @GraphicsMode(GraphicsMode.Mode.NATIVE)
 @Config(
     sdk = [33],
     qualifiers = "xxhdpi",
 )
-class BasicScreenshotTest {
+class BasicSnapshotTest {
 
     private val testDispatcher = UnconfinedTestDispatcher()
 
@@ -41,7 +41,7 @@ class BasicScreenshotTest {
     @OptIn(ExperimentalRoborazziApi::class, ExperimentalCoilApi::class)
     @Test
     fun testOverlay() = runTest(testDispatcher) {
-        val experienceResponse = TestJsonLoader.loadJsonFromAssetsDirectory("Screenshot", "Overlay.json")
+        val experienceResponse = TestJsonLoader.loadJsonFromAssetsDirectory("Snapshot", "Overlay.json")
 
         val engine = FakeImageLoaderEngine.Builder()
             .intercept(true, ColorDrawable(android.graphics.Color.RED))
