+++ roktux/src/test/java/com/rokt/roktux/snapshot/SnapshotTest.kt
@@ -1,0 +1,3 @@
+package com.rokt.roktux.snapshot
+
+interface SnapshotTest
